Major catchup and cleanup commit
Since last commit, a large new refactorization has been embarked on.
The new code is in the js2/ folder, mostly with filenames ending in 3.js.
It is more or less a new build from the ground up, although the core
xwd.js (now xwd3.js) is largely unchanged - the .pos field of the xwdCell
class is done differently being the main difference.
We now have a heirarchy of classes, with Crossword from xwd3.js
subclassed to xwdInterface in xwdInterface3.js - which adds abstract
aspects of interface, mostly cursor movement - and that further sublassed
to xwdInterfaceHtml which includes the _actual_ front-end html stuff.
The subclassing modules tend to tack some stuff on to the original,
and also use watcher.js, which allows fields of existing objects or
classes to be turned into pseudo-properties with a list of functions
to call when the property is altered. So the html presentation layer
actually 'listens' for changes to the underlying objects. This means
no need to add extra calls to display updates etc. in core xwd code.