New buttons for navigation and local storage, all with dynamic (responsive) layout.
Somewhat akin to the old serialize function, the main structure now has a pseudo-property
called content, being a string of characters for the content of the cells ('.' for none).
This makes the save and load pretty trivial. We could set up cookies as virtual local storage,
but for now if window.localStorage is not there, we use a new { } so it's only session long
anyway - which should really be drawn to user's attention.