Changed from 'confirm' dialog to modified button
Now using same approach as Guardian - change style of button (and add
"CONFIRM"), for up to 3 seconds. Second click actually calls the
relevant function. This still applies to any function with "ALL" in it's
description. If we ever internationalise the software, this will have to
be modified to be language-independent. But for now the interface is
only being used for English-language crosswords, so not yet an issue.