First efforts toward mobile-friendliness
The master module (Master3.js) now checks whether the URL has "?mobile"
added, and if so uses xwdInterfaceHtmlMobile.js instead of
xwdInterfaceHtml.js.  Much of the code still doesn't need to be changed
(e.g. reading the puzzle data), so this is still not quite right.
Initially the only difference is trying out the "contenteditable"
attribute, applied to the div element in each of the grid's cells. This
should bring up the virtual keyboard on a virtual device, but I may then
have to remove my own keyboard listening and instead monitor the
content, which ther browser will allow the user to change directly.