Change to single hidden dummy input
Instead of making each cell visibly content-editable, we give the focus
to a text input box hidden behind the grid. There was a bit of fiddling
arounbd to do to make sure we intercept printable characters (in case of
complications if the text box fills up) while allowing special
characters to be processed normally - e.g. F5 for refresh, PgUp & PgDn -
anything we're not using and won't print into the input box.